+++ src/Database.hpp
@@ -54,7 +54,7 @@ public:
 	return buffers[i];
     }
 
-    unsigned long size() { return buffers.size(); }
+    unsigned int size() { return buffers.size(); }
 
 private:
     std::vector<AudioBuffer<T> > buffers;
